refactor(SPV-1087): add review suggestions.
