+++ internal/api/v1/user/configs/configstest/response_200_status_code.json
